Guild & Gathering Changes
Added a guild recruitment section to help find nearby players that're not in a guild. Also allows for quick whispering by adding them to your clipboard, can have requirements on search and will not show the same person again if whispered.

Modified it so gathering fame amount shows up in the gathering tracking section.