Added Killboard
Implemented a killboard section that can take in a player name, guild name, or alliance name. It will track for any kills or deaths everyone minute based on the supplied name. Alliance will always override a Guild, a Guild will always override a Player. Supplies basic information when clicking on each event also provides a url to the official web url for that given event for additional information.